+++ lib/src/ServiceProvider/Sender/VerificationCodeSender.dart
@@ -1,0 +1,6 @@
+import 'package:interactivesso_datatypes/interactivesso_datatypes.dart';
+
+abstract class VerificationCodeSender{
+  Future<void> sendVerificationCode(VerificationCode code, UserInfo user, AvatarInfo? userAvatar, [String? locale]);
+  Future<void> sendOAuthVerificationCode(OAuthVeriCode code, UserInfo user, ApplicationInfo app, AvatarInfo? userAvatar, AvatarInfo? appAvatar, [String? locale]);
+}