+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.0.3
 
-- 
+- Add Avatar Factories
+- Add Captcha Service
+- Add Vericode Related Content Provider and Sender Abstractions.
 
 ## 0.0.2